Bootcamp , courses Ownership and create controll

+++ controllers/bootcamps.js
@@ -37,6 +37,17 @@ export const getBootcamp = asyncHandler(async (req, res, next) => {
 // @access private
 export const createBootcamps = asyncHandler(async (req, res, next) => {
 
+    // Add user to req.body
+    req.body.user = req.user.id;
+
+    // Check for publisher bootcamp
+    const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+    //If the user is not an admin, they can onyl add one bootcamp
+    if (publishedBootcamp && req.user.role !== "admin") {
+        return next(new ErrorResponse(`The user with ID ${req.user.id} has already published a bootcamp`));
+    }
+
     const bootcamp = await Bootcamp.create(req.body);
     res.status(201).json({
         success: true,
@@ -51,10 +62,18 @@ export const createBootcamps = asyncHandler(async (req, res, next) => {
 // @access Private
 export const updateBootcamps = asyncHandler(async (req, res, next) => {
 
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+    let bootcamp = await Bootcamp.findById(req.params.id);
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
     }
+
+    // Make sure user is Bootcamp Owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401));
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.user.id, req.body, { new: true, runValidators: true });
+
     res.status(200).json({ success: true, data: bootcamp })
 })
 
@@ -66,6 +85,11 @@ export const deleteBootcamps = asyncHandler(async (req, res, next) => {
     const bootcamp = await Bootcamp.findById(req.params.id);
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
+    }
+
+    // Make sure user is Bootcamp Owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete this bootcamp`, 401));
     }
 
     await bootcamp.deleteOne();
@@ -84,6 +108,11 @@ export const bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     // BootcampId validation
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
+    }
+
+    // Make sure user is Bootcamp Owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401));
     }
 
     // File validation
