+++ controllers/courses.js
@@ -53,11 +53,16 @@ export const getCourse = asyncHandler(async (req, res, next) => {
 export const addCourse = asyncHandler(async (req, res, next) => {
 
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
     if (!bootcamp) {
         return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`, 404));
+    }
+    // Make sure user is Bootcamp Owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`, 401));
     }
     const course = await Course.create(req.body);
 
@@ -79,6 +84,11 @@ export const updateCourse = asyncHandler(async (req, res, next) => {
 
     if (!course) {
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404);
+    }
+
+    // Make sure user is Course Owner
+    if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}`, 401));
     }
 
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -104,6 +114,10 @@ export const deleteCourse = asyncHandler(async (req, res, next) => {
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404);
     }
 
+    // Make sure user is Course Owner
+    if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`, 401));
+    }
     await Course.deleteOne();
 
     res.status(200).json({
